Decompose CommandLine god class into focused command classes (#404)
- Create Commands::Base with shared infrastructure (with_connection, run_hook)
- Extract 11 command classes: Pull, Push, Diff, Add, Rm, Mv, Addlocale,
  Rmlocale, Init, Match, Status
- Create Runner as thin dispatcher with COMMAND_MAP
- Add CommandLine = Runner alias for backward compatibility
- Delete command_line.rb
