+++ .rubocop_todo.yml
@@ -16,7 +16,8 @@ Lint/DuplicateMethods:
 # AllowedMethods: call
 Naming/PredicateMethod:
   Exclude:
-    - 'lib/web_translate_it/command_line.rb'
+    - 'lib/web_translate_it/runner.rb'
+    - 'lib/web_translate_it/commands/diff.rb'
 
 # Offense count: 13
 # Configuration parameters: CountAsOne.
