+++ bin/wti
@@ -135,7 +135,7 @@ end
 
 begin
   WebTranslateIt::Connection.turn_debug_on if command_options.debug
-  WebTranslateIt::CommandLine.new(command, command_options, global_options, ARGV, File.expand_path('.'))
+  WebTranslateIt::Runner.new(command, command_options, global_options, ARGV, File.expand_path('.'))
 rescue Interrupt
   puts StringUtil.failure("\nQuitting...")
   exit 1
