+++ lib/web_translate_it.rb
@@ -24,7 +24,19 @@ require 'web_translate_it/translation'
 require 'web_translate_it/term'
 require 'web_translate_it/term_translation'
 require 'web_translate_it/auto_fetch'
-require 'web_translate_it/command_line'
+require 'web_translate_it/commands/base'
+require 'web_translate_it/commands/pull'
+require 'web_translate_it/commands/push'
+require 'web_translate_it/commands/diff'
+require 'web_translate_it/commands/add'
+require 'web_translate_it/commands/rm'
+require 'web_translate_it/commands/mv'
+require 'web_translate_it/commands/addlocale'
+require 'web_translate_it/commands/rmlocale'
+require 'web_translate_it/commands/init'
+require 'web_translate_it/commands/match'
+require 'web_translate_it/commands/status'
+require 'web_translate_it/runner'
 require 'web_translate_it/project'
 
 module WebTranslateIt
